# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "operator",
        "version": "1.0",
        "description": "オペレーターサービス"
    },
    "servers": [
        {
            "url": "/operator"
        }
    ],
    "paths": {
...
        "/session": {
            "post": {
                "tags": [
                    "セッション確認"
                ],
                "description": "セッションIDを受け、有効か確認します。",
                "requestBody": {
                    "description": "セッションID",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "セッション確認のリクエストボディ",
                                "properties": {
                                    "sessionId": {
                                        "type": "string",
                                        "description": "種別"
                                    },
                                    "extendFlg": {
                                        "type": "boolean",
                                        "description": "SESSION延長フラグ"
                                    }
                                },
                                "example": {
                                    "sessionId": "013f51805a2611ab89400f80e75686bca965f9f6519640a99fb2d4180556968f",
                                    "extendFlg": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "セッション確認に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loginResponse"
                                },
                                "example": {
                                    "sessionId": "013f51805a2611ab89400f80e75686bca965f9f6519640a99fb2d4180556968f",
                                    "operatorId": 1,
                                    "type": 3,
                                    "loginId": "root_member01",
                                    "name": "運営メンバー01",
                                    "mobilePhone": "09011112222",
                                    "auth": {
                                        "member": {
                                            "add": true,
                                            "update": true,
                                            "delete": true
                                        }
                                    },
                                    "lastLoginAt": "2019-10-18 14:07:43.111",
                                    "passwordChangedFlg": true,
                                    "attributes": {},
                                    "roles": [
                                        {
                                            "_value": 1,
                                            "_ver": 1
                                        }
                                    ],
                                    "block": {
                                        "_value": 10001,
                                        "_ver": 1
                                    },
                                    "actor": {
                                        "_value": 10002,
                                        "_ver": 1
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "セッション確認で無効なセッションIDだった際のレスポンスコード"
                    }
                }
            }
        },
        "/ind/session": {
            "post": {
                "tags": [
                    "個人用セッション確認"
                ],
                "description": "セッションIDを受け、有効か確認します。",
                "requestBody": {
                    "description": "セッションID",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "セッション確認のリクエストボディ",
                                "properties": {
                                    "sessionId": {
                                        "type": "string",
                                        "description": "種別"
                                    },
                                    "extendFlg": {
                                        "type": "boolean",
                                        "description": "SESSION延長フラグ"
                                    }
                                },
                                "example": {
                                    "sessionId": "013f51805a2611ab89400f80e75686bca965f9f6519640a99fb2d4180556968f",
                                    "extendFlg": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "セッション確認に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loginResponse"
                                },
                                "example": {
                                    "sessionId": "013f51805a2611ab89400f80e75686bca965f9f6519640a99fb2d4180556968f",
                                    "operatorId": 1,
                                    "type": 0,
                                    "loginId": "ind_01",
                                    "pxrId": "ind_01",
                                    "name": "個人01",
                                    "mobilePhone": "09011112222",
                                    "lastLoginAt": "2019-10-18 14:07:43.111",
                                    "passwordChangedFlg": true,
                                    "attributes": {},
                                    "block": {
                                        "_value": 10001,
                                        "_ver": 1
                                    },
                                    "actor": {
                                        "_value": 10002,
                                        "_ver": 1
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "セッション確認で無効なセッションIDだった際のレスポンスコード"
                    }
                }
            }
        },
        "/password/{operatorId}": {
            "put": {
                "tags": [
                    "パスワードリセット"
                ],
                "description": "オペレーターのパスワードをリセットします。",
                "parameters": [
                    {
                        "in": "header",
                        "name": "session",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "セッション情報",
                            "example": {
                                "sessionId": "a1da98a9a3f146b7418ddb5267bb241b0c603c61c8102f3e1e25f08aee375223",
                                "operatorId": 1,
                                "type": 3,
                                "loginId": "mng_member",
                                "name": "mng_member",
                                "auth": {
                                    "member": {
                                        "add": true,
                                        "update": true,
                                        "delete": true
                                    }
                                },
                                "lastLoginAt": "2020-01-01 00:00:00.000+0900",
                                "attributes": {},
                                "block": {
                                    "_value": 1000110,
                                    "_ver": 1
                                },
                                "actor": {
                                    "_value": 1000001,
                                    "_ver": 1
                                }
                            }
                        }
                    },
                    {
                        "name": "operatorId",
                        "description": "オペレーターID",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "新パスワード",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "パスワードリセットのリクエストボディ",
                                "properties": {
                                    "newHpassword": {
                                        "type": "string",
                                        "description": "新しいハッシュパスワード"
                                    }
                                },
                                "example": {
                                    "newHpassword": "cd08374a2f7d9827ac1e592e497bf0268b0d2af4a84e1d55c690348012202948"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "パスワードリセットに成功した際のレスポンスコード",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "_code": {
                "type": "object",
                "properties": {
                    "_value": {
                        "type": "integer",
                        "description": "カタログコード"
                    },
                    "_ver": {
                        "type": "integer",
                        "description": "カタログバージョン"
                    }
                }
            },
            "auth": {
                "type": "object",
                "description": "権限",
                "properties": {
                    "member": {
                        "type": "object",
                        "description": "オペレーター操作権限",
                        "properties": {
                            "add": {
                                "type": "boolean",
                                "description": "追加権限"
                            },
                            "update": {
                                "type": "boolean",
                                "description": "更新権限"
                            },
                            "delete": {
                                "type": "boolean",
                                "description": "削除権限"
                            }
                        }
                    }
                }
            },
...
            "loginResponse": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "string",
                        "description": "セッションID"
                    },
                    "operatorId": {
                        "type": "integer",
                        "description": "オペレーターID"
                    },
                    "type": {
                        "type": "integer",
                        "description": "種別"
                    },
                    "loginId": {
                        "type": "string",
                        "description": "ログインID"
                    },
                    "pxrId": {
                        "type": "string",
                        "description": "PXR-ID"
                    },
                    "name": {
                        "type": "string",
                        "description": "表示名"
                    },
                    "mobilePhone": {
                        "type": "string",
                        "description": "携帯電話番号"
                    },
                    "auth": {
                        "$ref": "#/components/schemas/auth"
                    },
                    "lastLoginAt": {
                        "type": "string",
                        "description": "前回ログイン日時"
                    },
                    "passwordChangedFlg": {
                        "type": "boolean",
                        "description": "パスワード変更フラグ"
                    },
                    "attributes": {
                        "type": "object",
                        "description": "属性"
                    },
                    "roles": {
                        "type": "array",
                        "description": "ロール設定",
                        "items": {
                            "$ref": "#/components/schemas/_code"
                        }
                    },
                    "block": {
                        "$ref": "#/components/schemas/_code"
                    },
                    "actor": {
                        "$ref": "#/components/schemas/_code"
                    }
                }
            },
...
        }
    }
}